Refactor Crockford's base32
